fix(database): Ajusta tipos de colunas e adiciona role
- Altera o tipo das colunas de chave estrangeira para `unsignedBigInteger` nas migrations.
- Remove a migration `add_codigo_pessoa_to_users_table`.
- Adiciona a role `Operador` no `RoleSeeder`.

+++ database/migrations/2025_10_14_165401_create_lancamentos_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
                   ->references('codigo_pessoa')
                   ->on('pessoas')
                   ->onDelete('cascade');
-            $table->integer('usuario_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id')
                   ->references('id')
                   ->on('users')
