+++ database/migrations/2025_10_15_144145_create_grupo_usuario_table.php
@@ -12,12 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('grupo_usuario', function (Blueprint $table) {
-            $table->integer('usuario_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id')
                   ->references('id')
                   ->on('users')
                   ->onDelete('cascade');
-            $table->integer('grupo_id');
+            $table->unsignedBigInteger('grupo_id');
             $table->foreign('grupo_id')
                   ->references('id')
                   ->on('grupos')
